% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_functions.R
\name{calgaro_2020_16S_gingival_healthy}
\alias{calgaro_2020_16S_gingival_healthy}
\title{Calgaro 2020 - 16S (subgingival vs supragingival plaque)}
\usage{
calgaro_2020_16S_gingival_healthy()
}
\value{
A TreeSummarizedExperiment.
}
\description{
\code{calgaro_2020_16S} imports the 16S data used in Calgaro 2020 to
compare subgingival vs supragingival plaque samples. These samples are a
subset taken from the HMP16SData package (v1.2.0).
}
\details{
\strong{Biological signal:} anaerobic and aerobic bacteria are under- and over-
represented in supragingival plaque samples.
}
\section{Reference}{

Calgaro, M., Romualdi, C., Waldron, L. et al. Assessment of statistical methods from single
cell, bulk RNA-seq, and metagenomics applied to microbiome data. Genome Biol 21, 191 (2020).
https://doi.org/10.1186/s13059-020-02104-1
}

\keyword{internal}
