% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataset.R
\name{.assembleTreeSummarizedExperiment}
\alias{.assembleTreeSummarizedExperiment}
\title{Assemble TreeSummarizedExperiment}
\usage{
.assembleTreeSummarizedExperiment(x, dat_url)
}
\arguments{
\item{dat_url}{A character string with the URL from Zenodo.}

\item{dat_name}{A character string with the name of the dataset.}
}
\value{
A TreeSummarizedExperiment
}
\description{
\code{.assembleTreeSummarizedExperiment} assembles a TreeSummarizedDataset
taking as input the name of the dataset and the URL. This is a helper
function for the \code{\link{getDataset}} function.
}
\keyword{internal}
