% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_functions.R
\name{.beghini2019Nychanesmicrobiome}
\alias{.beghini2019Nychanesmicrobiome}
\title{Dataset used Beghini, 2019 (nychanesmicrobiome)}
\usage{
.beghini2019Nychanesmicrobiome()
}
\value{
A TreeSummarizedExperiment object
}
\description{
\code{.beghini2019Nychanesmicrobiome} imports the dataset used in the nychanes paper.
A good portion of this code was taken directly from the github repository
of the \code{nychnamesmicrobiome} package at
\url{https://github.com/waldronlab/nychanesmicrobiome/blob/master/R/loadQiimeData.R}.
}
\details{
Reference:
Beghini, F., Renson, A., Zolnik, C. P., Geistlinger, L., Usyk, M., Moody,
T. U., ... & Waldron, L. (2019). Tobacco exposure associated with oral
microbiota oxygen utilization in the New York City Health and Nutrition
Examination Study. Annals of epidemiology, 34, 18-25.
}
\keyword{internal}
