% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_functions.R
\name{.Beghini_2019_16S_smoking}
\alias{.Beghini_2019_16S_smoking}
\title{Beghini 2019 (smoking)}
\usage{
.Beghini_2019_16S_smoking()
}
\value{
A TreeSummarizedExperiment object.
}
\description{
\code{beghini_2019_16S} imports the dataset used in the nychanes paper.
A good portion of this code was taken directly from the github repository
of the \code{nychnamesmicrobiome} package at
\url{https://github.com/waldronlab/nychanesmicrobiome/blob/master/R/loadQiimeData.R}.
}
\section{Reference}{

Beghini, F., Renson, A., Zolnik, C. P., Geistlinger, L., Usyk, M., Moody,
T. U., ... & Waldron, L. (2019). Tobacco exposure associated with oral
microbiota oxygen utilization in the New York City Health and Nutrition
Examination Study. Annals of epidemiology, 34, 18-25.
}

\keyword{internal}
