% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_functions.R
\name{.calgaro2020Datasets}
\alias{.calgaro2020Datasets}
\title{Datasets Used in Calgaro, 2020}
\usage{
.calgaro2020Datasets(x)
}
\arguments{
\item{x}{A character vector of length 1. Valid options: 16S_Stool_TongueDorsum, 16S_Gingiva_Mucosa,
16S_Subgingival_Supragingival, WMS_Stool_TongueDorsum, WMS_Schizophrenia, WMS_CRC.}
}
\value{
A TreeSummarizedExperiment.
}
\description{
\code{.calgaro2020Datasets} imports one of the six datasets used in the paper of Calgaro, 2020
for benchmarking of methods for differential abundance (DA) analysis on microbiome data. Three
of the datasets were generated with 16S rRNA sequencing and come from the V35 dataset in the
\code{\link{HMP16SData}} package (v1.2.0). The other three datasets were generated with Whole
Metagenome Sequencing and come from the HMP_2012 (Stool_TongueDorsum), Castro-NallarE_2015
(Schizophrenia), and ZellerG_2014 (CRC) datasets in the curatedMetagenomicData (v1.12.3)
package.
}
\details{
Reference:
Calgaro, M., Romualdi, C., Waldron, L. et al. Assessment of statistical methods from single
cell, bulk RNA-seq, and metagenomics applied to microbiome data. Genome Biol 21, 191 (2020).
https://doi.org/10.1186/s13059-020-02104-1
}
\examples{

\dontrun{

library(MicrobiomeBenchmarkData)
library(SummarizedExperiment)

x <- .calgaro2020Datasets(x = "16S_Subgingival_Supragingival")
table(colData(x)$HMP_BODY_SUBSITE)

y <- .calgaro2020Datasets(x = "WMS_Stool_TongueDorsum")
table(colData(y)$body_subsite)

}

}
\keyword{internal}
