% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biocfilecache.R
\name{.getCache}
\alias{.getCache}
\title{Get MicrobiomeBenchmarkData's Cache}
\usage{
.getCache()
}
\description{
\code{.getCache} creates or loads a cache to store files downloaded through the
\code{MicrobiomeBenchmarkData} package.
}
\keyword{internal}
