% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biocfilecache.R
\name{.getResource}
\alias{.getResource}
\title{Get resource}
\usage{
.getResource(dat_name, dat_url)
}
\arguments{
\item{dat_name}{A character string with the name of the dataset.}

\item{dat_url}{A character string with the URL from Zenodo.}
}
\value{
A character string containing the path to the count matrix in the
cache.
}
\description{
\code{.getResource} downloads the count matrix and store it in the cache.
}
\keyword{internal}
