% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biocfilecache.R
\name{.getResourceFromCache}
\alias{.getResourceFromCache}
\title{Get Resource From Cache}
\usage{
.getResourceFromCache(fpath)
}
\arguments{
\item{fpath}{A character vector of length 1 indicating the full path to the resource that must be downloaded.}
}
\value{
A character vector of length 1 containing the full path to the resource in the cache.
}
\description{
\code{getResourceFromCache} gets the path to a resource in the cache generated with
\code{\link{.getCache}}.
}
\details{
\code{.getResourceFromCache} first searches for a resource in the cache. If the resource is not
found, then the resource is downloaded from the web to the cache.
}
\keyword{internal}
