% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicrobiomeBenchmarkData.R
\name{.getResourcePath}
\alias{.getResourcePath}
\title{Get resource path}
\usage{
.getResourcePath(resource, suffix)
}
\arguments{
\item{resource_name}{A character string with the name of the dataset.}

\item{resource_url}{A character string with the URL from Zenodo.}
}
\value{
A character string containing the path to the count matrix in the
cache.
}
\description{
\code{.getResource} downloads the count matrix and store it in the cache.
}
\keyword{internal}
