% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biocfilecache.R
\name{.getResourcePathFromCache}
\alias{.getResourcePathFromCache}
\title{Get resource path from the cache}
\usage{
.getResourcePathFromCache(resource_name)
}
\arguments{
\item{resource_name}{A character string indicating the name of a resource.}
}
\value{
A character string indicating the full path to the resource in the cache.
}
\description{
\code{getResourcePathFromCache} gets the path to a resource in the cache generated with
\code{\link{.getCache}}.
}
\keyword{internal}
