% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicrobiomeBenchmarkData.R
\name{.getSampleMetadata}
\alias{.getSampleMetadata}
\title{Get sample metadata}
\usage{
.getSampleMetadata()
}
\value{
A data frame with sample metadata.
}
\description{
\code{.getSampleMetadata} returns sampleMetadata.
}
\keyword{internal}
