% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.hmp16SDataCalgaro2020}
\alias{.hmp16SDataCalgaro2020}
\title{HMP16S Data from Calgaro 2020}
\usage{
.hmp16SDataCalgaro2020(x)
}
\arguments{
\item{x}{A character vector of length 1; the name of the dataset to be imported.}
}
\value{
A TreeSummarizedExperiment
}
\description{
\code{.hmp16SDataCalgaro2020} imports the HMP16S filtered data from Calgaro, 2020,  which was
used for enrichment analysis in that paper.
}
\keyword{internal}
