% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.hmp16SDataIntersects}
\alias{.hmp16SDataIntersects}
\title{HMP16SData Intersects}
\usage{
.hmp16SDataIntersects(subsites)
}
\arguments{
\item{subsites}{A character vector of length 2; the names of the body subsites.}
}
\value{
A list of common samples and subjects between the V13 and V35 datasets.
}
\description{
\code{.hmp16SDataIntersects} creates a list of common samples and subjects of the V13 and V35
datasets from the \code{HMP16SData} package based on two body subsites.
}
