% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.hmp16SDataSubset}
\alias{.hmp16SDataSubset}
\title{Subset HMP16SData}
\usage{
.hmp16SDataSubset(se, intersect)
}
\arguments{
\item{se}{A SummarizedExperiment imported with V13 or V35 functions.}

\item{intersect}{A list of intersects of samples and subjects (internal data of the \code{MicrobiomeBenchmarkData}).}
}
\value{
A TreeSummarizedExperiment.
}
\description{
\code{.hmp16SDataSubset} subsets a dataset, either V13 or V35, from the \code{\link{HMP16SData}}
based on common samples and subjects between both datasets.
}
\keyword{internal}
