% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_functions.R
\name{.HMP_2012_16S_gingival_a}
\alias{.HMP_2012_16S_gingival_a}
\title{.HMP_2012_16S_gingival_a}
\usage{
.HMP_2012_16S_gingival_a()
}
\value{
A TreeSummarizedExperiment.
}
\description{
\code{.HMP_2012_16S_gingival_a} imports the 16S data used in Calgaro 2020 to
compare subgingival vs supragingival plaque samples. These samples are a
subset taken from the HMP16SData package (v1.2.0). Variable region 3-5.
}
\keyword{internal}
