% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_functions.R
\name{.HMP_2012_16S_gingival_b}
\alias{.HMP_2012_16S_gingival_b}
\title{.HMP_2012_16S_gingival_b}
\usage{
.HMP_2012_16S_gingival_b()
}
\value{
A TreeSummarizedExperiment.
}
\description{
\code{.HMP_2012_16S_gingival_b} imports subgingival and supragingival samples
from the HMP 2012 dataset from subjects that are present in both the V13 and
V35 datasets from the HMP16SData package.
}
\keyword{internal}
