% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_functions.R
\name{.HMP_2012_WMS_gingival}
\alias{.HMP_2012_WMS_gingival}
\title{.HMP_2012_WMS_gingival}
\usage{
.HMP_2012_WMS_gingival()
}
\value{
A TreeSummarizedExperiment.
}
\description{
\code{.HMP_2012_WMS_gingival} imports the WGS samples used in Calgaro 2020
to compare subgingival vs supragingival plaque. This is an updated dataset
taken from curatedMetagenomicData v3 instead of curatedMetagenomicData v1.
}
\keyword{internal}
