% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.intersectSubjects}
\alias{.intersectSubjects}
\title{Intersect Subjects}
\usage{
.intersectSubjects(x, subsite1, subsite2, visit = 1)
}
\arguments{
\item{x}{A SummarizedExperiment imported by the V13 of V35 functions.#'}

\item{subsite1}{A character vector of length 1.}

\item{subsite2}{A character vector of length 1.}

\item{visit}{An integer; visit number.}
}
\value{
Common subjects between the V13 and V35 datasets.
}
\description{
\code{.intersectSubjects} creates a vector with the intersect of subjects present in
two body subsites of an HMP16SData (imported with V13 or V35).
}
