% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDatasets.R
\name{.listDatasets}
\alias{.listDatasets}
\title{List datasets}
\usage{
.listDatasets()
}
\value{
A tibble with available datasets.
}
\description{
\code{.listDatasets} displays the list of available datasets.
}
\keyword{internal}
