% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.sampleSEBySex}
\alias{.sampleSEBySex}
\title{Sample By Sex}
\usage{
.sampleSEBySex(x)
}
\arguments{
\item{x}{A SummarizedExperiment from the HMP16SData package}
}
\value{
A SummarizedExperiment with equal number of males and females in samples.
}
\description{
\code{.sampleBySex} is a helper function to randomly select equal number of males and females per
HMP_BODY_SUBSITE. It implements \code{set.seed} to obtain the same random samples. Currently,
this only works with datasets from the MP16SData package.
}
