% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicrobiomeBenchmarkData.R
\name{getDataset}
\alias{getDataset}
\title{Get dataset}
\usage{
getDataset(x, dryrun = TRUE)
}
\arguments{
\item{x}{A character vector with the name(s) of the dataset(s). If empty
and dryrun = TRUE, it returns a message with the names of the available
datasets. If empty and dryrun = FALSE, it returns a list of
TreeSummarizedExperiments with all of the datasets.}

\item{dryrun}{If TRUE, only returns a message and invisibly returns the
names of the datasets as a character vector. If FALSE, it returns the
TreeSummarizedExperiment datasets indicated in the argument 'x'.}
}
\value{
A list of TreeSummarizedExperiments.
}
\description{
\code{getDataset} imports datasets as TreeSummarizedExperiment objects.
}
\examples{

datasets_names <- getDataset()
datasets <- getDataset("HMP_2012_16S_gingival_V35_subset", dryrun = FALSE)

}
