% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataSet.R
\name{getDataSets}
\alias{getDataSets}
\title{Get Data Sets}
\usage{
getDataSets(x)
}
\arguments{
\item{x}{A numeric vector containing the indexes of the datasets to be loaded.}
}
\value{
A list of TreeSummarizedExperiments.
}
\description{
\code{getDataSets} returns a list containing datasets that have been specified by index according
to the output of \code{\link{dataSets}}.
}
\examples{

library(MicrobiomeBenchmarkData)
datasets <- getDataSets(x = 1)
datasets

}
