% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_functions.R
\name{HMP_16S_V35_gingival_healthy}
\alias{HMP_16S_V35_gingival_healthy}
\title{HMP16SData V35 - healthy gingival plaque}
\usage{
HMP_16S_V35_gingival_healthy()
}
\value{
A TreeSummarizedExperiment.
}
\description{
\code{HMP_16S_V35_gingival_healthy}
}
\examples{

library(MicrobiomeBenchmarkData)
x <- HMP_16S_V35_gingival_healthy()

}
