% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_functions.R
\name{HMP_WMS_cMD3_gingival_healthy}
\alias{HMP_WMS_cMD3_gingival_healthy}
\title{Calgaro 2020 - WGS cMD3 (subgingival vs supragingival plaque)}
\usage{
HMP_WMS_cMD3_gingival_healthy()
}
\value{
A TreeSummarizedExperiment.
}
\description{
\code{calgaro_2020_WGS_cMD3} imports the WGS samples used in Calgaro 2020
to compare subgingival vs supragingival plaque. This is an updated dataset
taken from curatedMetagenomicData v3 instead of curatedMetagenomicData v1.
}
\details{
\strong{Biological signal:} anaerobic and aerobic bacteria are under- and over-
represented in supragingival plaque samples.
}
\section{Reference}{

Calgaro, M., Romualdi, C., Waldron, L. et al. Assessment of statistical methods from single
cell, bulk RNA-seq, and metagenomics applied to microbiome data. Genome Biol 21, 191 (2020).
https://doi.org/10.1186/s13059-020-02104-1.
}

\examples{

library(MicrobiomeBenchmarkData)
x <- HMP_WMS_cMD3_gingival_healthy

}
