% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataets.R
\name{listDatasets}
\alias{listDatasets}
\title{List datasets}
\usage{
listDatasets()
}
\value{
A tibble with available datasets.
}
\description{
\code{listDatasets} displyas the list of available datasets.
}
\examples{

library(MicrobiomeBenchmarkData)
listDatasets()

}
\seealso{
\code{\link{getDatasets}}
}
