% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicrobiomeBenchmarkData.R
\docType{package}
\name{MicrobiomeBenchmarkDataAnalyses}
\alias{MicrobiomeBenchmarkDataAnalyses}
\title{MicrobiomeBenchmarkData}
\description{
The \code{MicrobiomeBenchmarkData} provide functions for accessing
various datasets.
}
