% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicrobiomeBenchmarkData.R
\name{removeCache}
\alias{removeCache}
\title{Remove cache}
\usage{
removeCache(ask = interactive())
}
\arguments{
\item{ask}{If TRUE, a prompt will appear asking the user to confirm removal
of cache. Default value is given by the \code{interactive} function.}
}
\value{
NULL The cache and all of its contents are removed.
}
\description{
\code{removeCache} removes all files saved in the cache.
}
\examples{

## Remove cache
removeCache()

}
