% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleMetadata.R
\docType{data}
\name{sampleMetadata}
\alias{sampleMetadata}
\title{sampleMetadata}
\format{
A data.frame.
}
\usage{
sampleMetadata
}
\description{
A data frame with the combined metadata of all of the samples in the
datasets provided through the MicrobiomeBenchmarkData package.
}
\keyword{datasets}
