% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataets.R
\name{showDatasets}
\alias{showDatasets}
\title{Show Datasets}
\usage{
showDatasets()
}
\value{
A tibble with available datasets.
}
\description{
\code{showDatasets} shows the list of available datasets in the \code{MicrobiomeBenchmarkData}
package.
}
\examples{

library(MicrobiomeBenchmarkData)
showDatasets()

}
\seealso{
\code{\link{getDatasets}}
}
