% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_functions.R
\name{spikeInBacteria}
\alias{spikeInBacteria}
\title{Spike-in bacteria dataset}
\usage{
spikeInBacteria()
}
\value{
A TreeSummarizedExperiment
}
\description{
\code{spikeInBacteria} provides a dataset with spike-in bacteria data.
TODO.
}
\section{Reference}{

Stämmler, F., Gläsner, J., Hiergeist, A. et al. Adjusting microbiome
profiles for differences in microbial load by spike-in bacteria.
Microbiome 4, 28 (2016). https://doi.org/10.1186/s40168-016-0175-0
}

\keyword{internal}
