% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_functions.R
\name{stoolVsTongueDorsumDataset1}
\alias{stoolVsTongueDorsumDataset1}
\title{Stool vs Tongue Dorsum Dataset 1}
\usage{
stoolVsTongueDorsumDataset1()
}
\value{
A TreeSummarizedExperiment.
}
\description{
\code{stoolVsTongueDorsumDataset1} loads a TreeSummarizedExperiment of Stool vs
Tongue Dorsum from the HMP16S project used in Calgaro, 2020 for enrichment analysis.
}
