% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_functions.R
\name{stoolVsTongueDorsumDataset3}
\alias{stoolVsTongueDorsumDataset3}
\title{Stool vs Tongue Dorsum Dataset 3}
\usage{
stoolVsTongueDorsumDataset3()
}
\value{
A TreeSummarizedExperiment.
}
\description{
\code{stoolVsTongueDorsumDataset3} loads a TreeSummarizedExperiment of Stool vs Tongue Dorsum
from the V35 dataset from the \code{\link{HMP16SData}} package. The samples
have been filtered to only include subjects present in both the V13 and V35 datasets of the
HMP16SData package and include only the visit number 1 to the research center. Furthermore, the samples
include equal numbers of males and females.
}
