% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_functions.R
\name{stoolVsTongueDorsumDataset4}
\alias{stoolVsTongueDorsumDataset4}
\title{Stool vs Tongue Dorsum Dataset 4}
\usage{
stoolVsTongueDorsumDataset4()
}
\value{
A TreeSummarizedExperiment
}
\description{
\code{stoolVsTongueDorsumDataset4} imports stool and tongue dorsum samples from the HMP_2012
study in curatedMetagenomicData (a single sample per subject).
}
