% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_functions.R
\name{subgingivalVsSupragingivalDataset1}
\alias{subgingivalVsSupragingivalDataset1}
\title{Subgingival vs Supragingival Plaque Dataset 1}
\usage{
subgingivalVsSupragingivalDataset1()
}
\value{
A TreeSummarizedExperiment.
}
\description{
\code{subgingivalVsSupragingivalDataset1} loads a TreeSummarizedExperiment of Subgingival vs
Supragingival Plaque from the HMP16S project used in Calgaro, 2020 for enrichment analysis.
}
