% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_functions.R
\name{subgingivalVsSupragingivalDataset3}
\alias{subgingivalVsSupragingivalDataset3}
\title{Subgingival vs Supragingival Plaque Dataset 3}
\usage{
subgingivalVsSupragingivalDataset3()
}
\value{
A TreeSummarizedExperiment.
}
\description{
\code{subgingivalVsSupragingivalDataset3} loads a TreeSummarizedExperiment of Subgingival vs
Supragingival Plaque from the V35 dataset from the \code{\link{HMP16SData}} package. The samples
have been filtered to only include subjects present in both the V13 and V35 datasets of the
HMP16SData package and include only the visit number 1 to the research center. Furthermore, the samples
include equal numbers of males and females.
}
