% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataSet.R
\name{dataSets}
\alias{dataSets}
\title{Data Sets
\code{dataSets} returns the available datasets in the  \code{MicrobiomeBenchmarkData} package.}
\usage{
dataSets()
}
\value{
A tibble with all available datasets.
}
\description{
Data Sets
\code{dataSets} returns the available datasets in the  \code{MicrobiomeBenchmarkData} package.
}
\examples{

library(MicrobiomeBenchmarkData)
datasets <- dataSets()

}
