% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataets.R
\name{.datasets}
\alias{.datasets}
\title{Data Sets
\code{dataSets} imports the table containing the list of available datasets in the
\code{MicrobiomeBenchmarkData} package.}
\usage{
.datasets()
}
\value{
A tibble with all available datasets.
}
\description{
Data Sets
\code{dataSets} imports the table containing the list of available datasets in the
\code{MicrobiomeBenchmarkData} package.
}
\keyword{internal}
