% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biocfilecache.R
\name{.getResource}
\alias{.getResource}
\title{Get resource}
\usage{
.getResource(resource_name, FUN)
}
\arguments{
\item{resource_name}{A single character string with the name of the
dataset.}

\item{FUN}{Function to download the dataset.}
}
\value{
A TreeSummarizedExperiment.
}
\description{
\code{.getResource} loads a resource (dataset) from the cache.
If the resource does not exist in the cache, it will be downloaded first
from the internet with a helper function specified with the \code{FUN} argument.
}
\details{
The \code{.getResource} function is meant to be used inside the functions
used to download the datasets only, e.g.
\code{\link{.HMP_2012_16S_gingival_a}}.
}
\seealso{
\code{\link{.getCache}};
\code{\link{.getResourcePathFromCache}}
}
\keyword{internal}
