% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataets.R
\name{getDatasets}
\alias{getDatasets}
\title{Get datasets}
\usage{
getDatasets(index = NULL, name = NULL)
}
\arguments{
\item{index}{A numeric vector containing the indexes of the datasets to be loaded.}

\item{name}{A character vector containing the names of the datasets to be loaded.}
}
\value{
A list vector containing TreeSummarizedExperiments.
}
\description{
\code{getDataSets} imports one or several datasets as
TreeSummarizedExperiment(s) stored in a list vector. The datasets
can accessed according to their index or name as specified in the output of
the \code{\link{listDatasets}} function. The datasets can be loaded
according to index or name, but not both.
}
\examples{

library(MicrobiomeBenchmarkData)

## Load datasets by index
datasets1 <- getDatasets(index = c(1,3))
datasets1

## Load datasets by name
datasets_names <- c("calgaro_2020_16S_gingival_healthy",
                    "beghini_2019_16S_saliva_smoking" )
datasets2 <- getDatasets(name = datasets_names)
datasets2

}
\seealso{
\code{\link{listDatasets}}
}
