% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataets.R
\name{getDatasets}
\alias{getDatasets}
\title{Get datasets}
\usage{
getDatasets(x)
}
\arguments{
\item{x}{A numeric vector containing the indexes of the datasets to be loaded.}
}
\value{
A list of TreeSummarizedExperiments.
}
\description{
\code{getDataSets} returns a list containing datasets that have been specified by index according
to the output of \code{\link{showDatasets}}.
}
\examples{

library(MicrobiomeBenchmarkData)
datasets <- getDatasets(x = c(1,3))
datasets

}
\seealso{
\code{\link{showDatasets}}
}
