% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDatasets.R
\name{getDatasets}
\alias{getDatasets}
\title{Get datasets}
\usage{
getDatasets(dat_name = NULL)
}
\arguments{
\item{dat_name}{A character vector containing the names of the datasets to be loaded. If
'NULL' (default), a message indicating how to get detailed information
about the datasets is printed along with the names of available datasets;
furthermore, the list of available datasets is invisibly returned as a
character vector.}
}
\value{
A list containing TreeSummarizedExperiments. If \code{NULL}, a character vector
with available datasets (invisibly returned).
}
\description{
\code{getDataSets} imports one or several datasets as
TreeSummarizedExperiment(s) stored in a list vector. The datasets
can accessed according to their index or name as specified in the output of
the \code{\link{.listDatasets}} function. The datasets can be loaded
according to index or name, but not both.
}
\examples{

## Show datasets and save datasets names as a character vector
datasets <- getDatasets()

## Download datasets 1 and two in the list
x <- getDatasets(datasets[1:2])

length(x)
vapply(x, class, character(1))

}
