% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicrobiomeBenchmarkData-package.R
\docType{package}
\name{MicrobiomeBenchmarkData}
\alias{MicrobiomeBenchmarkData}
\title{MicrobiomeBenchmarkData}
\description{
The \code{MicrobiomeBenchmarkData} provide functions for accessing
various microbiome datasets with biological ground truth.
}
