% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biocfilecache.R
\name{removeCache}
\alias{removeCache}
\title{Remove cache}
\usage{
removeCache()
}
\description{
\code{removeCache} removes all files saved in the cache. It will ask for
confirmation before removing the cache.
}
\keyword{internal}
