% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicrobiomeBenchmarkData.R
\name{removeCache}
\alias{removeCache}
\title{Remove cache}
\usage{
removeCache(ask = interactive())
}
\value{
NULL The cache and all of its contents are removed.
}
\description{
\code{removeCache} removes all files saved in the cache. It will ask for
confirmation before removing the cache (if in an interactive session).
}
\examples{

## Remove cache
removeCache()

}
